+++ src/index.js
@@ -10,6 +10,7 @@ import '@/components/Basic/RestaurantCategoryIcon/RestaurantCategoryIcon';
 
 import '@/components/RestaurantList/RestaurantList';
 import '@/components/RestaurantList/RestaurantItem';
+import '@/components/NewRestaurantModal/NewRestaurantForm';
 import '@/components/NewRestaurantModal/NewRestaurantModal';
 import '@/components/OnOffButton';
 import '@/components/Tab';
